# Do not gather local all together at the beginning of typeck