--- a/tests/ui/methods/disambiguate-multiple-blanket-impl.stderr
+++ b/tests/ui/methods/disambiguate-multiple-blanket-impl.stderr
@@ -1,18 +1,3 @@
-error[E0223]: ambiguous associated type
-  --> $DIR/disambiguate-multiple-blanket-impl.rs:36:12
-   |
-LL |     let _: S::Type;
-   |            ^^^^^^^
-   |
-help: use fully-qualified syntax
-   |
-LL -     let _: S::Type;
-LL +     let _: <S as A>::Type;
-   |
-LL -     let _: S::Type;
-LL +     let _: <S as B>::Type;
-   |
-
 error[E0034]: multiple applicable items in scope
   --> $DIR/disambiguate-multiple-blanket-impl.rs:30:8
    |
@@ -61,6 +46,21 @@ LL +     <S as A>::CONST;
    |
 LL -     S::CONST;
 LL +     <S as B>::CONST;
+   |
+
+error[E0223]: ambiguous associated type
+  --> $DIR/disambiguate-multiple-blanket-impl.rs:36:12
+   |
+LL |     let _: S::Type;
+   |            ^^^^^^^
+   |
+help: use fully-qualified syntax
+   |
+LL -     let _: S::Type;
+LL +     let _: <S as A>::Type;
+   |
+LL -     let _: S::Type;
+LL +     let _: <S as B>::Type;
    |
 
 error: aborting due to 3 previous errors