… tests/ui/methods/disambiguate-multiple-trait-2.stderr
… tests/ui/stability-attribute/generics-default-stability.stderr
… tests/ui/stability-attribute/suggest-vec-allocator-api.stderr
… tests/ui/traits/bound/on-structs-and-enums-locals.stderr
… tests/ui/traits/bound/on-structs-and-enums-xc1.stderr