r? lcnr