@compiler-errors 2025-05-05
@bors r=lcnr rollup=never