@rust-log-analyzer 2025-05-06
A job failed! Check out the build log: [(web)](https://github.com/rust-lang/rust/actions/runs/14861197416/job/41726244905) [(plain)](https://github.com/rust-lang/rust/commit/f5d3fe273b8b9e7125bf8856d44793b6cc4b6735/checks/41726244905/logs)

<details><summary><i>Click to see the possible cause of the failure (guessed by this bot)</i></summary>

```plain
   Compiling diff v0.1.13
   Compiling glob-match v0.2.1
   Compiling citool v0.1.0 (/home/runner/work/rust/rust/src/ci/citool)
    Finished `release` profile [optimized] target(s) in 37.82s
     Running `target/release/citool post-merge-report 651e9cf327358b28db7e37a2ae61727f4a2ef232 f5d3fe273b8b9e7125bf8856d44793b6cc4b6735`
Downloading metrics of job aarch64-gnu
Downloading metrics of job aarch64-gnu-debug
Downloading metrics of job arm-android
Downloading metrics of job armhf-gnu
Downloading metrics of job dist-aarch64-linux
Downloading metrics of job dist-android
Downloading metrics of job dist-arm-linux
Downloading metrics of job dist-armhf-linux
Downloading metrics of job dist-armv7-linux
Downloading metrics of job dist-i586-gnu-i586-i686-musl
Downloading metrics of job dist-i686-linux
Downloading metrics of job dist-loongarch64-linux
Downloading metrics of job dist-loongarch64-musl
Downloading metrics of job dist-ohos
Downloading metrics of job dist-powerpc-linux
Downloading metrics of job dist-powerpc64-linux
Downloading metrics of job dist-powerpc64le-linux
Downloading metrics of job dist-riscv64-linux
Downloading metrics of job dist-s390x-linux
Downloading metrics of job dist-various-1
Downloading metrics of job dist-various-2
Downloading metrics of job dist-x86_64-freebsd
Downloading metrics of job dist-x86_64-illumos
Downloading metrics of job dist-x86_64-linux
Downloading metrics of job dist-x86_64-linux-alt
Downloading metrics of job dist-x86_64-musl
Downloading metrics of job dist-x86_64-netbsd
Downloading metrics of job i686-gnu-1
Downloading metrics of job i686-gnu-2
Downloading metrics of job i686-gnu-nopt-1
Downloading metrics of job i686-gnu-nopt-2
Downloading metrics of job mingw-check
Downloading metrics of job test-various
Downloading metrics of job x86_64-rust-for-linux
Downloading metrics of job x86_64-gnu
Downloading metrics of job x86_64-gnu-stable
Downloading metrics of job x86_64-gnu-aux
Downloading metrics of job x86_64-gnu-debug
Downloading metrics of job x86_64-gnu-distcheck
Downloading metrics of job x86_64-gnu-llvm-20-1
Downloading metrics of job x86_64-gnu-llvm-20-2
Downloading metrics of job x86_64-gnu-llvm-20-3
Downloading metrics of job x86_64-gnu-llvm-19-1
Downloading metrics of job x86_64-gnu-llvm-19-2
Downloading metrics of job x86_64-gnu-llvm-19-3
Downloading metrics of job x86_64-gnu-nopt
Downloading metrics of job x86_64-gnu-tools
Downloading metrics of job dist-x86_64-apple
Downloading metrics of job dist-apple-various
Downloading metrics of job x86_64-apple-1
Downloading metrics of job x86_64-apple-2
Downloading metrics of job dist-aarch64-apple
Downloading metrics of job aarch64-apple
Downloading metrics of job x86_64-msvc-1
Downloading metrics of job x86_64-msvc-2
Downloading metrics of job i686-msvc-1
Downloading metrics of job i686-msvc-2
Downloading metrics of job x86_64-msvc-ext1
Downloading metrics of job x86_64-msvc-ext2
Downloading metrics of job x86_64-msvc-ext3
Downloading metrics of job x86_64-mingw-1
Downloading metrics of job x86_64-mingw-2
Downloading metrics of job dist-x86_64-msvc
Downloading metrics of job dist-i686-msvc
Downloading metrics of job dist-aarch64-msvc
Downloading metrics of job dist-i686-mingw
Downloading metrics of job dist-x86_64-mingw
Downloading metrics of job dist-x86_64-msvc-alt
<details>
<summary>What is this?</summary>
This is an experimental post-merge analysis report that shows differences in test outcomes between the merged PR and its parent PR.
</details>


Comparing 651e9cf327358b28db7e37a2ae61727f4a2ef232 (parent) -> f5d3fe273b8b9e7125bf8856d44793b6cc4b6735 (this PR)

# Test differences
<details>
<summary>Show 7 test diffs
</summary>


7 doctest diffs were found. These are ignored, as they are noisy.
</details>

<details>
<summary>Test dashboard</summary>

Run

```bash
cargo run --manifest-path src/ci/citool/Cargo.toml -- \
    test-dashboard f5d3fe273b8b9e7125bf8856d44793b6cc4b6735 --output-dir test-dashboard
```
And then open `test-dashboard/index.html` in your browser to see an overview of all executed tests.

</details>

# Job duration changes
1. [aarch64-apple](https://github.com/rust-lang-ci/rust/actions/runs/14857380790#summary-41713904141): 4201.4s -> 3643.9s (-13.3%)
2. [dist-x86_64-apple](https://github.com/rust-lang-ci/rust/actions/runs/14857380790#summary-41713901672): 7884.2s -> 8887.1s (12.7%)
3. [x86_64-apple-1](https://github.com/rust-lang-ci/rust/actions/runs/14857380790#summary-41713902596): 8180.6s -> 9156.9s (11.9%)
4. [dist-x86_64-freebsd](https://github.com/rust-lang-ci/rust/actions/runs/14857380790#summary-41713890414): 4900.8s -> 5478.5s (11.8%)
5. [dist-apple-various](https://github.com/rust-lang-ci/rust/actions/runs/14857380790#summary-41713902169): 7467.3s -> 6786.8s (-9.1%)
6. [x86_64-gnu-llvm-20-3](https://github.com/rust-lang-ci/rust/actions/runs/14857380790#summary-41713899055): 7060.5s -> 6744.6s (-4.5%)
7. [dist-aarch64-apple](https://github.com/rust-lang-ci/rust/actions/runs/14857380790#summary-41713903561): 5424.5s -> 5665.6s (4.4%)
8. [x86_64-mingw-2](https://github.com/rust-lang-ci/rust/actions/runs/14857380790#summary-41713908099): 7233.5s -> 6919.6s (-4.3%)
9. [dist-s390x-linux](https://github.com/rust-lang-ci/rust/actions/runs/14857380790#summary-41713889144): 5479.2s -> 5245.3s (-4.3%)
10. [x86_64-msvc-ext2](https://github.com/rust-lang-ci/rust/actions/runs/14857380790#summary-41713906744): 5939.8s -> 5714.8s (-3.8%)

<details>
<summary>How to interpret the job duration changes?</summary>

Job durations can vary a lot, based on the actual runner instance
that executed the job, system noise, invalidated caches, etc. The table above is provided
mostly for t-infra members, for simpler debugging of potential CI slow-downs.
</details>

no pull requests found for branch "master"
##[error]Process completed with exit code 1.
Post job cleanup.
```

</details>