@lqd 2025-05-06
cc @kobzol on that post merge job failure