@Kobzol 2025-05-06
Thanks! https://github.com/rust-lang/rust/pull/140703 should hopefully help.