@nnethercote 2025-05-06
@compiler-errors: some amazing perf results there, well done. How did you find this/think of this?