@compiler-errors 2025-05-06
@nnethercote: Gathering locals at the beginning of typeck ended up leading to a hang in the new solver (`-Znext-solver`) for `tests/ui/issues/issue-29466.rs`. Investigating it led to my solution to register locals more closely to the first chance of them being used, which I only really intended to fix the hang in the new solver. 

I guess it coincidentally ended up also being positive perf-wise in the old solver, but I didn't look too closely into why. Probably has something to do with holding a ton of pending `_: Sized` goals in the fulfillment context, which get repeatedly checked if they're unstalled.