@panstromek 2025-05-12
perf triage:

Large improvements outweigh regresions. Regressions are spurious (all regressed crates returned back to previous state in a following commit).

@rustbot label: +perf-regression-triaged